+++ webpack/ForemanWebhooks/Routes/Webhooks/Components/WebhookTable/Components/Formatters/nameToEditFormatter.js
@@ -1,25 +1,32 @@
 import React from 'react';
-import NameToEditCell from '../NameToEditCell';
+import { Button } from '@patternfly/react-core';
 
-const nameToEditFormatter = (controllerPluralize, onClick) => (
-  value,
-  extra
-) => {
+const nameToEditFormatter = onClick => (value, extra) => {
   const row = extra?.rowData ?? value;
   const canEdit = row.canEdit ?? row.can_edit;
-  const { id, name } = row;
-  const label = extra?.rowData != null ? value : name;
+  const { id } = row;
+  const label = extra?.rowData != null ? value : row.name;
 
-  return (
-    <NameToEditCell
-      active={canEdit}
-      id={id}
-      name={encodeURI(name)}
-      controller={controllerPluralize}
-      onClick={onClick}
+  return canEdit ? (
+    <Button
+      ouiaId="name-edit-active-button"
+      variant="link"
+      isInline
+      component="span"
+      onClick={() => onClick(id)}
     >
       {label}
-    </NameToEditCell>
+    </Button>
+  ) : (
+    <Button
+      ouiaId="name-edit-disabled-button"
+      variant="link"
+      isInline
+      isDisabled
+      component="span"
+    >
+      {label}
+    </Button>
   );
 };
